Enhances codesniffer-run to detect and use a local phpcs.xml.
Adds a `-r Standard` | `-r rules.xml` command line option to manually specify the coding standard to use.

Also adds the ability to manually specify dirs/files to sniff instead of the defaults.

Combined, this should about make it a complete replacement for the default `phpcs` command with "better" defaults across the board.

Closes #67.